saas-backend/app/schemas/user.py
from pydantic import BaseModel, EmailStr


class LoginRequest(BaseModel):
    username: str
    password: str

class LoginResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"

class RegisterRequest(BaseModel):
    username: str
    password: str
    role: str

class User(BaseModel):
    role: str
    username: str
