saas-backend/requirements.txt
fastapi
uvicorn
permit
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
motor
pydantic-settings
